more mapping

+++ get_next_line_100/get_next_line.h
@@ -6,7 +6,7 @@
 /*   By: miandrad <miandrad@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/11/14 14:54:26 by miandrad          #+#    #+#             */
-/*   Updated: 2023/01/04 18:49:53 by miandrad         ###   ########.fr       */
+/*   Updated: 2023/01/05 13:10:50 by miandrad         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -23,7 +23,7 @@
 
 char	*get_next_line(int fd);
 char	*ft_strjoin(char const *str1, char const *str2);
-size_t	ft_strlen_gnl(const char *str, size_t n);
+size_t	ft_strlen_gnl(const char *str, int n);
 char	*ft_strchr(const char *str, int c);
 char	*ft_substr(char const *str, unsigned int start, size_t len);
 void	*ft_calloc(size_t noob, size_t size);
